# [wip] Return a frozen array of bytes when producing keypairs