--- a/nacl.js
+++ b/nacl.js
@@ -125,6 +125,28 @@ function crypto_core_hsalsa20(out,inp,k,c) {
   return 0;
 }
 
+function key_arrays_to_keypair(pk, sk) {
+  var pkArray, skArray;
+  return Object.defineProperties({}, {
+    publicKey: {
+      get: function() {
+        if (pkArray === undefined) {
+          pkArray = Object.freeze(Array.from(pk));
+        }
+        return pkArray;
+      }
+    },
+    secretKey: {
+      get: function() {
+        if (skArray === undefined) {
+          skArray = Object.freeze(Array.from(sk));
+        }
+        return skArray;
+      }
+    }
+  });
+}
+
 var sigma = new Uint8Array([101, 120, 112, 97, 110, 100, 32, 51, 50, 45, 98, 121, 116, 101, 32, 107]);
             // "expand 32-byte k"
 
@@ -948,8 +970,16 @@ function checkBoxLengths(pk, sk) {
 
 function checkArrayTypes() {
   for (var i = 0; i < arguments.length; i++) {
-    if (!(arguments[i] instanceof Uint8Array))
-      throw new TypeError('unexpected type, use Uint8Array');
+    if (arguments[i] instanceof Object && typeof arguments[i].length === 'number') {
+      for(var j = 0; j < arguments[i].length; j++) {
+        var value = arguments[i][j];
+        if (typeof value !== 'number' || value < 0 || value > 255) {
+          break;
+        }
+      }
+      continue;
+    }
+    throw new TypeError('unexpected type, expected an array-like object with numerical values between 0-255');
   }
 }
 
@@ -1034,7 +1064,7 @@ nacl.box.keyPair = function() {
   var pk = new Uint8Array(crypto_box_PUBLICKEYBYTES);
   var sk = new Uint8Array(crypto_box_SECRETKEYBYTES);
   crypto_box_keypair(pk, sk);
-  return {publicKey: pk, secretKey: sk};
+  return key_arrays_to_keypair(pk, sk);
 };
 
 nacl.box.keyPair.fromSecretKey = function(secretKey) {
@@ -1043,7 +1073,7 @@ nacl.box.keyPair.fromSecretKey = function(secretKey) {
     throw new Error('bad secret key size');
   var pk = new Uint8Array(crypto_box_PUBLICKEYBYTES);
   crypto_scalarmult_base(pk, secretKey);
-  return {publicKey: pk, secretKey: new Uint8Array(secretKey)};
+  return key_arrays_to_keypair(pk, new Uint8Array(secretKey));
 };
 
 nacl.box.publicKeyLength = crypto_box_PUBLICKEYBYTES;
@@ -1098,7 +1128,7 @@ nacl.sign.keyPair = function() {
   var pk = new Uint8Array(crypto_sign_PUBLICKEYBYTES);
   var sk = new Uint8Array(crypto_sign_SECRETKEYBYTES);
   crypto_sign_keypair(pk, sk);
-  return {publicKey: pk, secretKey: sk};
+  return key_arrays_to_keypair(pk, sk);
 };
 
 nacl.sign.keyPair.fromSecretKey = function(secretKey) {
@@ -1107,7 +1137,7 @@ nacl.sign.keyPair.fromSecretKey = function(secretKey) {
     throw new Error('bad secret key size');
   var pk = new Uint8Array(crypto_sign_PUBLICKEYBYTES);
   for (var i = 0; i < pk.length; i++) pk[i] = secretKey[32+i];
-  return {publicKey: pk, secretKey: new Uint8Array(secretKey)};
+  return key_arrays_to_keypair(pk, new Uint8Array(secretKey));
 };
 
 nacl.sign.keyPair.fromSeed = function(seed) {
@@ -1118,7 +1148,7 @@ nacl.sign.keyPair.fromSeed = function(seed) {
   var sk = new Uint8Array(crypto_sign_SECRETKEYBYTES);
   for (var i = 0; i < 32; i++) sk[i] = seed[i];
   crypto_sign_keypair(pk, sk, true);
-  return {publicKey: pk, secretKey: sk};
+  return key_arrays_to_keypair(pk, sk);
 };
 
 nacl.sign.publicKeyLength = crypto_sign_PUBLICKEYBYTES;