@paulmillr 2022-08-03
Important: It only partially addresses it, many cases are still left w/o solution.

Quite a few libraries re-calculate public keys. There is no issue in doing that, instead of half-solutions.